Restructure sequelize setup to use one common configuration.
Previously, database migrations were run using the DATABASE_URL, and the actual application was run using individual options. The application has a need to run with SSL enabled all of the time, which was not possible to do using the DATABASE_URL option sequelize uses. This abstracts the configuration so that both use a full configuration object instead of a DATABASE_URL.
