+++ apps/backend/db/database.ts
@@ -1,0 +1,5 @@
+import AppConfig from '../config/app_config';
+
+const appConfig = new AppConfig();
+
+module.exports = appConfig.getDbConfig();
