+++ apps/backend/package.json
@@ -9,7 +9,8 @@
     "build": "nest build",
     "lint": "eslint \"{src,migrations,seeders,test}/**/*.ts\" --fix",
     "lint:ci": "eslint \"{src,migrations,seeders,test}/**/*.ts\" --max-warnings 0",
-    "start": "node ../../dist/server/main",
+    "sequelize-cli": "ts-node node_modules/.bin/sequelize",
+    "start": "node ../../dist/server/src/main",
     "start:debug": "nest start --debug --watch",
     "start:dev": "nest start --watch",
     "test": "jest --silent",
