+++ cmd.sh
@@ -2,4 +2,4 @@
 set -e
 yarn backend sequelize-cli db:migrate
 yarn backend sequelize-cli db:seed:all
-node main.js
+yarn backend start
